Templated generic database editor. 泛用型模板化数据库编辑器。
With XLibEditor, you can define a template for your database first
(define what kind of data your database have, and structure) , then
create your database with this template , platform is mac, supporting
PLIST file format, and NSKeyedAchiver format for cocoa.

I made this editor because I want to create a RPG at beginning, but RPG
have mass of data, if use files like excel or txt to create the
database is not convenient. Different datas in different type, and also
equipment needs reference to skill, monster reference to item, photos,
etc. That’s really a trouble. so I tried to create a database editor
for the game, but after actually doing it, that’s not easy, especially
if I change the settings of the game. So I thought about that, why I
can’t create a database Editor for every game?

Then I come up with that, I can cut the editor into two parts:
definition of the database (template), and database instance. Also
simplify the process of template config. The result is XLibEditor. And
introduction of photo preview, multi-languguage support, list, etc.

这是一个先根据应用、游戏来制作对应的数据库模板（定义类的细节），然后根据模板来编辑数据库的软件，平台是Mac，支持通用的PLIST格式（即XML
）和mac、ios用的COCOA支持比较好的NSKeyedArchiver打包。

制作这个编辑器最初是因为我想制作一个RPG类的游戏，但是因为这种游戏数据量大，用EXCEL或者TXT之类的编辑器显然不是很适合，特别是不同的对象
属性需要不同的数据类型，还有装备引用技能、怪物引用道具、图片等等各种麻烦的引用特别容易出错。所以就想到根据游戏来开发一个数据编辑器，但是实现做起
来，发现开发麻烦，而且如果游戏设定发生了更改，改动编辑器就更是麻烦到极点，就想，有没有办法一个编辑器适合不同类型的游戏呢？

于是就想到了将数据库的定义（模板）和数据库的实例切开来，同时简化模板制作的想法。当时找了一下也没有找到类似功能的软件，就做成了这个程序（嘛，有的
话就当练习和多一个选择了~）。

也根据自己开发游戏的经验加入了很多辅助性的功能，比如刚才说的引用、图片预览、列表等。